**feat(api): add support for psycopg2-binary and redis**
This commit adds support for psycopg2-binary and redis in our project. It includes the necessary packages and dependencies, including async-timeout, hiredis, and ocsp.

The changes include:

* Adding psycopg2-binary as a package dependency
* Adding redis as a package dependency
* Updating the `AUTH_USER_MODEL` setting to `accounts.User`
* Adding the necessary files for psycopg2-binary and redis

This commit improves the functionality of our project by providing support for PostgreSQL and Redis databases. It also enables us to use the `psycopg2-binary` package for PostgreSQL connections and the `redis` package for Redis connections.